Moved function ac_flag_match_mask out of iris_engine namespace down into models
